+++ src/lib/validation/index.js
@@ -1,0 +1,12 @@
+/* @flow */
+export { validateAddItems } from './validate-add-items';
+export { validateUser } from './validate-user';
+export { validatePurchase } from './validate-purchase';
+export { validateRemoveItems } from './validate-remove-items';
+export {
+  setUserNormalizer,
+  addToCartNormalizer,
+  purchaseNormalizer,
+  setCartNormalizer,
+  removeFromCartNormalizer
+} from './legacy-input-normalizers';
